third_party/blink/web_tests/external/wpt/css/css-logical/animations/logical-shorthand-relative-prioritization-by-number-of-components.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Logical shorthands follow the usual prioritization based on number of component longhands
  assert_equals: expected "300px" but got "0px"
Harness: the test ran to completion.
